Fix: align Web DTO types with API response types and fix LicenseFee serialization
- ReferenceDataDto.cs: change 21 reference-code primary key fields from int/int?
  to short/short? to match API entity types (MakeDto, ModelDto, VehicleTypeDto,
  FuelTypeDto, ClassDto, UnitOfMeasureDto, LicenseTypeDto, LicenseFeeDto,
  LossTypeDto, ExtraCodeDto). DriverLicenceDto.licence_code stays int because
  driver_licence_type_id is int on the API side.

- LicenseFeeApiService.cs: rename ApiLicenseFeeResponse properties from snake_case
  (licence_fee_code, licence_description, licence_fee) to PascalCase (LicenceFeeCode,
  Description, Fee) to match the PascalCase JSON keys the LicenseFeeController emits.
  The old names caused 100% silent data loss on all three fields.

- ReportsTotalKm.razor: cast literal 0 to (short)0 in Select lambda failure returns
  so the inferred tuple element type stays short, keeping the classDescriptions
  Dictionary<short,string?> lookup compatible after class_code became short.

+++ src/Services/FIS.Web/Models/ReferenceDataDto.cs
@@ -4,7 +4,7 @@ namespace FIS.Web.Models;
 
 public record MakeDto
 {
-    public int make_code { get; set; }
+    public short make_code { get; set; }
     [Required]
     public string? make_name { get; set; }
     public DateTime? date_updated { get; set; }
@@ -12,29 +12,29 @@ public record MakeDto
 
 public record ModelDto
 {
-    public int model_code { get; set; }
+    public short model_code { get; set; }
     [Required]
     public string? model_name { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "Make is required.")]
-    public int? make_code { get; set; }
+    public short? make_code { get; set; }
     public string? make_name { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "Unit of measure is required.")]
-    public int? unit_of_measure_code { get; set; }
+    public short? unit_of_measure_code { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "License type is required.")]
-    public int? licence_code { get; set; }
+    public short? licence_code { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "Class is required.")]
-    public int? class_code { get; set; }
+    public short? class_code { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "Vehicle type is required.")]
-    public int? type_code { get; set; }
+    public short? type_code { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "Fuel type is required.")]
-    public int? fuel_type_code { get; set; }
+    public short? fuel_type_code { get; set; }
     [Range(1, int.MaxValue, ErrorMessage = "License fee is required.")]
-    public int? licence_fee_code { get; set; }
-    public int? maint_trigger_code { get; set; }
+    public short? licence_fee_code { get; set; }
+    public short? maint_trigger_code { get; set; }
     public string? engine_type { get; set; }
-    public int? engine_capacity { get; set; }
-    public int? rated_power { get; set; }
-    public int? fuel_tank_capacity { get; set; }
+    public short? engine_capacity { get; set; }
+    public short? rated_power { get; set; }
+    public short? fuel_tank_capacity { get; set; }
     public decimal? target_consumption { get; set; }
     public int? target_tyre_life { get; set; }
     public int? service_interval { get; set; }
@@ -46,14 +46,14 @@ public record ModelDto
 
 public record VehicleTypeDto
 {
-    public int type_code { get; set; }
+    public short type_code { get; set; }
     [Required]
     public string? type_name { get; set; }
 }
 
 public record FuelTypeDto
 {
-    public int fuel_type_code { get; set; }
+    public short fuel_type_code { get; set; }
     [Required]
     public string? fuel_type_name { get; set; }
     public decimal? rate_per_litre { get; set; }
@@ -61,14 +61,14 @@ public record FuelTypeDto
 
 public record ClassDto
 {
-    public int class_code { get; set; }
+    public short class_code { get; set; }
     [Required]
     public string? class_description { get; set; }
 }
 
 public record UnitOfMeasureDto
 {
-    public int unit_of_measure_code { get; set; }
+    public short unit_of_measure_code { get; set; }
     [Required]
     public string? unit_description { get; set; }
     public string? unit_abbreviation { get; set; }
@@ -77,7 +77,7 @@ public record UnitOfMeasureDto
 
 public record LicenseTypeDto
 {
-    public int licence_code { get; set; }
+    public short licence_code { get; set; }
     [Required]
     public string? licence_description { get; set; }
     public string? licence_category { get; set; }
@@ -85,7 +85,7 @@ public record LicenseTypeDto
 
 public record LicenseFeeDto
 {
-    public int licence_fee_code { get; set; }
+    public short licence_fee_code { get; set; }
     [Required]
     public string? licence_description { get; set; }
     public decimal? licence_fee { get; set; }
@@ -100,14 +100,14 @@ public record DriverLicenceDto
 
 public record ExtraCodeDto
 {
-    public int extra_code { get; set; }
+    public short extra_code { get; set; }
     [Required]
     public string? extra_description { get; set; }
 }
 
 public record LossTypeDto
 {
-    public int loss_code { get; set; }
+    public short loss_code { get; set; }
     [Required]
     public string? loss_description { get; set; }
 }
